Increase host buffer size
Currently the host buffer only has space for 7 characters, this means
on systems with many scsi hosts (1000 or more), multipath will overflow
this buffer. This can happen pretty easily if there are a large number
of iscsi devices.  This patch increases the host buffer to hold 15
characters, which can deal with 100000000000000 scsi hosts.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>