multipathd/main.c: delete redundant judgment before LOG_MSG()
Problem:the LOG_MSG() doesn't note the path info "path offline" when the
path is in offline with the pp->offline = 1.

Reasons:There is no need to judge the strlen(b) again before calling
LOG_MSG(a, b) because there exists the judgmnet of strlen(b) in
defination of LOG_MSG(a,b).