allow multipath to retry removes of in-use devices
Occasionally, a multipath device is temporarily opened by things like
udev. This can cause the multipath flush commands to fail. While it is
possible to simply rerun the command, it can be very annoying for
scripts that are working with multipath devices. To deal with that,
it is now possible to tell multipath to retry failed removes. Either
running multipath with "-R <num>" or setting "remove_retries <num>"
in /etc/multipath.conf will make multipath retry failed removes
the specified number of times, with a 1 second delay between tries.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>