add "max_sectors_kb" config parameter
This patch adds the "max_sectors_kb" parameter to all multipath.conf
sections. Setting this will cause multipath to set the max_sectors_kb
queue parameter to the specified value on all of the paths before
creating the multipath device, which will inherit the value from the
paths.

If max_sectors_kb is out of sync between the multipath device and the
path devices, it's possible for multpiath to send down a request that is
too large for the path to handle.  When devices are discovered,
multipathd automatically sets up multipath devices on top of them. LVM
and filesystems can get mounted on top of that.  This means that users
who what to modify max_sectors_kb manually, may have to do it after
these devices are already in use.  This config option lets them simply
have multipath set up the values before creating the device.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>